Move fs utils from extra/fs.rs to keys.rs
